feat(studio): refine model catalog and goal composer
