+++ frontend/scripts/prepare-next-build.mjs
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+import { rmSync } from "node:fs";
+import { resolve } from "node:path";
+
+rmSync(resolve(import.meta.dirname, "..", ".next"), { recursive: true, force: true });
