+++ frontend/src/features/agent/ui/agent-composer-actions.tsx
@@ -53,7 +53,7 @@ export function AgentComposerActions({
   const activeIconClass = "bg-(--active) text-(--fg)";
 
   return (
-    <div className="agent-composer-actions-row flex min-h-9 items-center gap-0.5 bg-transparent px-2.5 pb-2 pt-0 text-xs">
+    <div className="agent-composer-actions-row flex min-h-9 items-center gap-0.5 bg-transparent px-2 pb-2.5 pt-0 text-xs">
       <input
         ref={fileInputRef}
         type="file"
