+++ frontend/src/features/agent/ui/git-diff-drawer.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { lazy, Suspense } from "react";
-import { Drawer, DrawerBody, DrawerHeader } from "@/ui/drawer";
+import { Drawer, DrawerBody, DrawerHeader, DrawerOverlay } from "@/ui/drawer";
 import type { GitSummary } from "@/features/agent/projects/types";
 
 const LazyGitDiffPanel = lazy(() =>
@@ -25,39 +25,30 @@ export function GitDiffDrawer({
   onClose: () => void;
 }) {
   return (
-    <div
-      className="fixed inset-0 z-50 flex justify-end bg-black/30"
-      onClick={onClose}
-      onKeyDown={(event) => {
-        if (event.key === "Escape") onClose();
-      }}
-      role="presentation"
-    >
-      <div className="flex h-full" onClick={(event) => event.stopPropagation()} role="dialog">
-        <Drawer width={720} className="h-full">
-          <DrawerHeader
-            title={gitBranch ? `Changes · ${gitBranch}` : "Changes"}
-            badge={
-              gitSummary?.isRepo ? (
-                <span className="inline-flex shrink-0 items-center gap-1 font-mono text-[length:var(--fs-xs)] tabular-nums">
-                  <span className="text-(--ok)">+{gitSummary.additions}</span>
-                  <span className="text-(--err)">-{gitSummary.deletions}</span>
-                </span>
-              ) : null
+    <DrawerOverlay onClose={onClose}>
+      <Drawer width={720} className="h-full">
+        <DrawerHeader
+          title={gitBranch ? `Changes · ${gitBranch}` : "Changes"}
+          badge={
+            gitSummary?.isRepo ? (
+              <span className="inline-flex shrink-0 items-center gap-1 font-mono text-[length:var(--fs-xs)] tabular-nums">
+                <span className="text-(--ok)">+{gitSummary.additions}</span>
+                <span className="text-(--err)">-{gitSummary.deletions}</span>
+              </span>
+            ) : null
+          }
+          onClose={onClose}
+        />
+        <DrawerBody className="p-0">
+          <Suspense
+            fallback={
+              <div className="p-4 text-[length:var(--fs-sm)] text-(--dim)">Loading diff…</div>
             }
-            onClose={onClose}
-          />
-          <DrawerBody className="p-0">
-            <Suspense
-              fallback={
-                <div className="p-4 text-[length:var(--fs-sm)] text-(--dim)">Loading diff…</div>
-              }
-            >
-              <LazyGitDiffPanel cwd={cwd} />
-            </Suspense>
-          </DrawerBody>
-        </Drawer>
-      </div>
-    </div>
+          >
+            <LazyGitDiffPanel cwd={cwd} />
+          </Suspense>
+        </DrawerBody>
+      </Drawer>
+    </DrawerOverlay>
   );
 }
