+++ frontend/src/features/agent/ui/session-goal-bar.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useCallback, useState } from "react";
-import { ChevronDown, CircleDot, Pause, Play, Trash2 } from "@/ui/icon-registry";
+import { ChevronDown, FilePenLine, Pause, Play, Save, Target, Trash2, X } from "@/ui/icon-registry";
 import { useMountSubscription } from "@/hooks/use-mount-subscription";
 
 type GoalStatus = "active" | "paused" | "blocked" | "complete" | "budget_limited";
@@ -51,6 +51,8 @@ export function SessionGoalBar({
 }) {
   const [goal, setGoal] = useState<SessionGoal | null>(null);
   const [expanded, setExpanded] = useState(false);
+  const [editing, setEditing] = useState(false);
+  const [draft, setDraft] = useState("");
 
   useMountSubscription(() => {
     let cancelled = false;
@@ -74,6 +76,7 @@ export function SessionGoalBar({
       if (body === null) {
         await fetch(url, { method: "DELETE" }).catch(() => undefined);
         setGoal(null);
+        setEditing(false);
         return;
       }
       await fetch(url, {
@@ -88,47 +91,72 @@ export function SessionGoalBar({
 
   if (!goal) return null;
   const paused = goal.status === "paused";
-  const terminal = goal.status === "complete" || goal.status === "blocked";
+  const terminal =
+    goal.status === "complete" || goal.status === "blocked" || goal.status === "budget_limited";
   const detailId = `session-goal-detail-${piSessionId}`;
+  const iconButtonClass =
+    "inline-flex h-7 w-7 shrink-0 items-center justify-center rounded-full text-(--fg)/42 transition-colors hover:bg-(--hover) hover:text-(--fg)/82 focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-(--fg)/25";
+
+  const startEditing = () => {
+    setDraft(goal.objective);
+    setEditing(true);
+    setExpanded(true);
+  };
+
+  const saveObjective = async () => {
+    const objective = draft.trim();
+    if (!objective) return;
+    await mutate({ objective });
+    setEditing(false);
+  };
 
   return (
     <section
       data-testid="session-goal-bar"
-      className="mx-auto mb-2 w-full max-w-[var(--composer-w)] overflow-hidden rounded-[24px] border border-(--border) bg-(--fg)/[0.025] text-[length:var(--fs-sm)] shadow-[0_8px_24px_rgba(0,0,0,0.08)]"
+      className="relative z-[90] mx-auto -mb-4 w-[calc(90%_-_26px)] max-w-[calc(var(--composer-w)*0.9_-_26px)] overflow-hidden rounded-[20px] border border-(--border)/80 bg-(--composer)/70 pb-4 text-[length:var(--fs-sm)] shadow-[0_8px_24px_rgba(0,0,0,0.08)] backdrop-blur-sm [corner-shape:superellipse(1.5)]"
     >
-      <div className="flex min-h-12 items-center gap-2 px-3.5 py-2">
-        <CircleDot
+      <div className="flex h-11 items-center gap-2 px-3">
+        <Target
           className={`h-4 w-4 shrink-0 ${
             goal.status === "active"
-              ? "text-(--ok,--accent)"
+              ? "text-(--fg)/56"
               : goal.status === "blocked"
                 ? "text-(--err)"
-                : "text-(--hl3)"
+                : "text-(--fg)/34"
           }`}
         />
         <button
           type="button"
           onClick={() => setExpanded((value) => !value)}
-          className="flex min-w-0 flex-1 items-center gap-2 text-left"
+          className="flex min-w-0 flex-1 items-center gap-2 text-left focus-visible:outline-none"
           aria-expanded={expanded}
           aria-controls={detailId}
         >
-          <span className="shrink-0 font-medium text-(--fg)/88">{STATUS_LABEL[goal.status]}</span>
+          <span className="shrink-0 font-medium text-(--fg)/82">{STATUS_LABEL[goal.status]}</span>
           {!expanded ? (
-            <span className="min-w-0 truncate text-(--fg)/48" title={goal.objective}>
+            <span className="min-w-0 truncate text-(--fg)/40" title={goal.objective}>
               {goal.objective}
             </span>
           ) : null}
         </button>
-        <span className="shrink-0 font-mono text-[length:var(--fs-sm)] tabular-nums text-(--fg)/48">
+        <span className="shrink-0 tabular-nums text-(--fg)/40">
           {formatElapsed(goal.createdAt)}
           {goal.turnBudget ? ` · ${goal.turnsUsed}/${goal.turnBudget}` : ""}
         </span>
+        <button
+          type="button"
+          onClick={startEditing}
+          className={iconButtonClass}
+          aria-label="Edit goal"
+          title="Edit goal"
+        >
+          <FilePenLine className="h-3.5 w-3.5" />
+        </button>
         {!terminal ? (
           <button
             type="button"
             onClick={() => void mutate({ status: paused ? "active" : "paused" })}
-            className="shrink-0 rounded-full p-1.5 text-(--fg)/55 transition-colors hover:bg-(--hover) hover:text-(--fg)"
+            className={iconButtonClass}
             aria-label={paused ? "Resume goal" : "Pause goal"}
             title={paused ? "Resume goal" : "Pause goal"}
           >
@@ -138,7 +166,7 @@ export function SessionGoalBar({
         <button
           type="button"
           onClick={() => void mutate(null)}
-          className="shrink-0 rounded-full p-1.5 text-(--fg)/55 transition-colors hover:bg-(--hover) hover:text-(--fg)"
+          className={iconButtonClass}
           aria-label="Clear goal"
           title="Clear goal"
         >
@@ -147,7 +175,7 @@ export function SessionGoalBar({
         <button
           type="button"
           onClick={() => setExpanded((value) => !value)}
-          className="shrink-0 rounded-full p-1.5 text-(--fg)/55 transition-colors hover:bg-(--hover) hover:text-(--fg)"
+          className={iconButtonClass}
           aria-label={expanded ? "Collapse goal" : "Expand goal"}
           title={expanded ? "Collapse goal" : "Expand goal"}
         >
@@ -157,12 +185,52 @@ export function SessionGoalBar({
         </button>
       </div>
       {expanded ? (
-        <p
-          id={detailId}
-          className="border-t border-(--border)/75 px-4 py-3 text-[length:var(--fs-base)] leading-relaxed text-(--fg)/62"
-        >
-          {goal.objective}
-        </p>
+        <div id={detailId} className="px-4 pb-4 pt-1">
+          {editing ? (
+            <div className="rounded-xl border border-(--border) bg-(--composer)/45 p-2.5">
+              <textarea
+                value={draft}
+                onChange={(event) => setDraft(event.target.value)}
+                onKeyDown={(event) => {
+                  if (event.key === "Escape") setEditing(false);
+                  if (event.key === "Enter" && (event.metaKey || event.ctrlKey)) {
+                    event.preventDefault();
+                    void saveObjective();
+                  }
+                }}
+                rows={2}
+                autoFocus
+                className="max-h-28 min-h-14 w-full resize-none bg-transparent text-[length:var(--fs-base)] leading-relaxed text-(--fg)/72 outline-none placeholder:text-(--fg)/30"
+                aria-label="Goal objective"
+              />
+              <div className="flex justify-end gap-1 pt-1">
+                <button
+                  type="button"
+                  onClick={() => setEditing(false)}
+                  className={iconButtonClass}
+                  aria-label="Cancel editing goal"
+                  title="Cancel"
+                >
+                  <X className="h-3.5 w-3.5" />
+                </button>
+                <button
+                  type="button"
+                  onClick={() => void saveObjective()}
+                  disabled={!draft.trim()}
+                  className={`${iconButtonClass} bg-(--fg)/90 text-(--bg) hover:bg-(--fg) hover:text-(--bg) disabled:opacity-35`}
+                  aria-label="Save goal"
+                  title="Save goal"
+                >
+                  <Save className="h-3.5 w-3.5" />
+                </button>
+              </div>
+            </div>
+          ) : (
+            <p className="max-w-[92%] text-[length:var(--fs-base)] leading-[1.55] text-(--fg)/48">
+              {goal.objective}
+            </p>
+          )}
+        </div>
       ) : null}
     </section>
   );
