+++ frontend/src/lib/types.ts
@@ -115,16 +115,7 @@ export interface LaunchProgressData extends LaunchProgress {
   message: string;
 }
 
-// --- Model discovery + recommendation ---
-
-export interface ModelRecommendation {
-  id: string;
-  name: string;
-  size_gb: number | null;
-  min_vram_gb: number | null;
-  description: string;
-  tags: string[];
-}
+// --- Model discovery + starter presets ---
 
 export interface StarterPreset {
   id: string;
